add DTO,MAPPER,ADMIN for creat materiaux,and controller for see full materiaux

+++ src/main/java/com/simplon/recyclascore/repositories/IMateriauxRepository.java
@@ -4,6 +4,8 @@ import com.simplon.recyclascore.models.Materiaux;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.util.Optional;
+
 @Repository
-public interface IMateriauxRepository extends JpaRepository<Materiaux, Long> {
+public interface IMateriauxRepository extends JpaRepository<Materiaux, Integer> {
 }
