+++ src/main/java/com/simplon/recyclascore/repositories/IMonoMaterialRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IMonoMaterialRepository extends JpaRepository<MonoMaterial, Long> {
+public interface IMonoMaterialRepository extends JpaRepository<MonoMaterial, Integer> {
     Iterable<MonoMaterial> findAllByMaterial(EnumMaterial material);
 }
