+++ src/main/java/com/simplon/recyclascore/repositories/IValidationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface IValidationRepository extends JpaRepository<Validation, Long> {
+public interface IValidationRepository extends JpaRepository<Validation, Integer> {
   Optional<Validation> findByCode(String code);
 }
